+++ packages/template-retail-react-app/app/components/_app/index.jsx
@@ -10,6 +10,7 @@ import PropTypes from 'prop-types'
 import {useLocation} from 'react-router-dom'
 import {getAssetUrl} from 'pwa-kit-react-sdk/ssr/universal/utils'
 import {getAppOrigin} from 'pwa-kit-react-sdk/utils/url'
+import StorefrontPreview from 'pwa-kit-react-sdk/storefront-preview'
 
 // Chakra
 import {Box, useDisclosure, useStyleConfig} from '@chakra-ui/react'
@@ -208,145 +209,151 @@ const App = (props) => {
 
     return (
         <Box className="sf-app" {...styles.container}>
-            <IntlProvider
-                onError={(err) => {
-                    if (err.code === 'MISSING_TRANSLATION') {
-                        // NOTE: Remove the console error for missing translations during development,
-                        // as we knew translations would be added later.
-                        console.warn('Missing translation', err.message)
-                        return
-                    }
-                    throw err
-                }}
-                locale={targetLocale}
-                messages={messages}
-                // For react-intl, the _default locale_ refers to the locale that the inline `defaultMessage`s are written for.
-                // NOTE: if you update this value, please also update the following npm scripts in `template-retail-react-app/package.json`:
-                // - "extract-default-translations"
-                // - "compile-translations:pseudo"
-                defaultLocale={DEFAULT_LOCALE}
-            >
-                <CategoriesProvider treeRoot={allCategories} locale={targetLocale}>
-                    <CurrencyProvider currency={currency}>
-                        <AmplienceContextProvider {...ampProps} showVse={showVse}>
-                            {showVse && <Toolbar {...ampProps} showVse={showVse} />}
-                            <RealtimeVisualization.Provider value={{ampVizSdk, status}}>
-                                <Seo>
-                                    <meta name="theme-color" content={THEME_COLOR} />
-                                    <meta
-                                        name="apple-mobile-web-app-title"
-                                        content={DEFAULT_SITE_TITLE}
-                                    />
-                                    <link
-                                        rel="apple-touch-icon"
-                                        href={getAssetUrl('static/img/global/apple-touch-icon.png')}
-                                    />
-                                    <link
-                                        rel="manifest"
-                                        href={getAssetUrl('static/manifest.json')}
-                                    />
+            {/* For pwa-kit v2, we need to use this getToken prop to make sure it can work properly
+                with Storefront Preview feature*/}
+            <StorefrontPreview getToken={() => window.localStorage.getItem('token')}>
+                <IntlProvider
+                    onError={(err) => {
+                        if (err.code === 'MISSING_TRANSLATION') {
+                            // NOTE: Remove the console error for missing translations during development,
+                            // as we knew translations would be added later.
+                            console.warn('Missing translation', err.message)
+                            return
+                        }
+                        throw err
+                    }}
+                    locale={targetLocale}
+                    messages={messages}
+                    // For react-intl, the _default locale_ refers to the locale that the inline `defaultMessage`s are written for.
+                    // NOTE: if you update this value, please also update the following npm scripts in `template-retail-react-app/package.json`:
+                    // - "extract-default-translations"
+                    // - "compile-translations:pseudo"
+                    defaultLocale={DEFAULT_LOCALE}
+                >
+                    <CategoriesProvider treeRoot={allCategories} locale={targetLocale}>
+                        <CurrencyProvider currency={currency}>
+                            <AmplienceContextProvider {...ampProps} showVse={showVse}>
+                                {showVse && <Toolbar {...ampProps} showVse={showVse} />}
+                                <RealtimeVisualization.Provider value={{ampVizSdk, status}}>
+                                    <Seo>
+                                        <meta name="theme-color" content={THEME_COLOR} />
+                                        <meta
+                                            name="apple-mobile-web-app-title"
+                                            content={DEFAULT_SITE_TITLE}
+                                        />
+                                        <link
+                                            rel="apple-touch-icon"
+                                            href={getAssetUrl(
+                                                'static/img/global/apple-touch-icon.png'
+                                            )}
+                                        />
+                                        <link
+                                            rel="manifest"
+                                            href={getAssetUrl('static/manifest.json')}
+                                        />
 
-                                    {/* Urls for all localized versions of this page (including current page)
+                                        {/* Urls for all localized versions of this page (including current page)
                             For more details on hrefLang, see https://developers.google.com/search/docs/advanced/crawling/localized-versions */}
-                                    {site.l10n?.supportedLocales.map((locale) => (
+                                        {site.l10n?.supportedLocales.map((locale) => (
+                                            <link
+                                                rel="alternate"
+                                                hrefLang={locale.id.toLowerCase()}
+                                                href={`${appOrigin}${buildUrl(location.pathname)}`}
+                                                key={locale.id}
+                                            />
+                                        ))}
+                                        {/* A general locale as fallback. For example: "en" if default locale is "en-GB" */}
                                         <link
                                             rel="alternate"
-                                            hrefLang={locale.id.toLowerCase()}
+                                            hrefLang={site.l10n.defaultLocale.slice(0, 2)}
                                             href={`${appOrigin}${buildUrl(location.pathname)}`}
-                                            key={locale.id}
                                         />
-                                    ))}
-                                    {/* A general locale as fallback. For example: "en" if default locale is "en-GB" */}
-                                    <link
-                                        rel="alternate"
-                                        hrefLang={site.l10n.defaultLocale.slice(0, 2)}
-                                        href={`${appOrigin}${buildUrl(location.pathname)}`}
-                                    />
-                                    {/* A wider fallback for user locales that the app does not support */}
-                                    <link
-                                        rel="alternate"
-                                        hrefLang="x-default"
-                                        href={`${appOrigin}/`}
-                                    />
-                                </Seo>
+                                        {/* A wider fallback for user locales that the app does not support */}
+                                        <link
+                                            rel="alternate"
+                                            hrefLang="x-default"
+                                            href={`${appOrigin}/`}
+                                        />
+                                    </Seo>
 
-                                <ScrollToTop />
+                                    <ScrollToTop />
 
-                                <Box id="app" display="flex" flexDirection="column" flex={1}>
-                                    <SkipNavLink zIndex="skipLink">Skip to Content</SkipNavLink>
+                                    <Box id="app" display="flex" flexDirection="column" flex={1}>
+                                        <SkipNavLink zIndex="skipLink">Skip to Content</SkipNavLink>
 
-                                    <Box {...styles.headerWrapper}>
-                                        {!isCheckout ? (
-                                            <Header
-                                                onMenuClick={onOpen}
-                                                onLogoClick={onLogoClick}
-                                                onMyCartClick={onCartClick}
-                                                onMyAccountClick={onAccountClick}
-                                                onWishlistClick={onWishlistClick}
-                                                logo={headerNav.icon}
-                                                logoAlt={headerNav.alt}
+                                        <Box {...styles.headerWrapper}>
+                                            {!isCheckout ? (
+                                                <Header
+                                                    onMenuClick={onOpen}
+                                                    onLogoClick={onLogoClick}
+                                                    onMyCartClick={onCartClick}
+                                                    onMyAccountClick={onAccountClick}
+                                                    onWishlistClick={onWishlistClick}
+                                                    logo={headerNav.icon}
+                                                    logoAlt={headerNav.alt}
+                                                >
+                                                    <HideOnDesktop>
+                                                        <DrawerMenu
+                                                            isOpen={isOpen}
+                                                            onClose={onClose}
+                                                            onLogoClick={onLogoClick}
+                                                            root={headerNav}
+                                                            footer={footerNav}
+                                                            locale={locale}
+                                                            logo={headerNav.icon}
+                                                            showVse={showVse}
+                                                        />
+                                                        {/* Need to fix */}
+                                                    </HideOnDesktop>
+
+                                                    <HideOnMobile>
+                                                        <AmplienceListMenu root={headerNav} />
+                                                    </HideOnMobile>
+                                                </Header>
+                                            ) : (
+                                                <CheckoutHeader />
+                                            )}
+                                        </Box>
+
+                                        {!isOnline && <OfflineBanner />}
+                                        <AddToCartModalProvider>
+                                            <SkipNavContent
+                                                style={{
+                                                    display: 'flex',
+                                                    flexDirection: 'column',
+                                                    flex: 1,
+                                                    outline: 0
+                                                }}
                                             >
-                                                <HideOnDesktop>
-                                                    <DrawerMenu
-                                                        isOpen={isOpen}
-                                                        onClose={onClose}
-                                                        onLogoClick={onLogoClick}
-                                                        root={headerNav}
-                                                        footer={footerNav}
-                                                        locale={locale}
-                                                        logo={headerNav.icon}
-                                                        showVse={showVse}
-                                                    />
-                                                    {/* Need to fix */}
-                                                </HideOnDesktop>
+                                                <Box
+                                                    as="main"
+                                                    id="app-main"
+                                                    role="main"
+                                                    display="flex"
+                                                    flexDirection="column"
+                                                    flex="1"
+                                                >
+                                                    <OfflineBoundary isOnline={false}>
+                                                        {children}
+                                                    </OfflineBoundary>
+                                                </Box>
+                                            </SkipNavContent>
 
-                                                <HideOnMobile>
-                                                    <AmplienceListMenu root={headerNav} />
-                                                </HideOnMobile>
-                                            </Header>
-                                        ) : (
-                                            <CheckoutHeader />
-                                        )}
+                                            {!isCheckout ? (
+                                                <Footer root={footerNav} />
+                                            ) : (
+                                                <CheckoutFooter />
+                                            )}
+
+                                            <AuthModal {...authModal} />
+                                        </AddToCartModalProvider>
                                     </Box>
-
-                                    {!isOnline && <OfflineBanner />}
-                                    <AddToCartModalProvider>
-                                        <SkipNavContent
-                                            style={{
-                                                display: 'flex',
-                                                flexDirection: 'column',
-                                                flex: 1,
-                                                outline: 0
-                                            }}
-                                        >
-                                            <Box
-                                                as="main"
-                                                id="app-main"
-                                                role="main"
-                                                display="flex"
-                                                flexDirection="column"
-                                                flex="1"
-                                            >
-                                                <OfflineBoundary isOnline={false}>
-                                                    {children}
-                                                </OfflineBoundary>
-                                            </Box>
-                                        </SkipNavContent>
-
-                                        {!isCheckout ? (
-                                            <Footer root={footerNav} />
-                                        ) : (
-                                            <CheckoutFooter />
-                                        )}
-
-                                        <AuthModal {...authModal} />
-                                    </AddToCartModalProvider>
-                                </Box>
-                            </RealtimeVisualization.Provider>
-                        </AmplienceContextProvider>
-                    </CurrencyProvider>
-                </CategoriesProvider>
-            </IntlProvider>
+                                </RealtimeVisualization.Provider>
+                            </AmplienceContextProvider>
+                        </CurrencyProvider>
+                    </CategoriesProvider>
+                </IntlProvider>
+            </StorefrontPreview>
         </Box>
     )
 }